The project contains multiple typos that cause confusions to other developers:
In README.md the document specifies that an environemnt variable 'APISECRET' needs to be passed in.

However the code actually reads the variable 'APISECET' (without R) which is a typo.
The consequence is that we cannot run the commands specified within the documentation.

Moreover the error is misleading. Once the apiSecret is not found, the exception 'API key is missing' is thrown.
'API secret is missing' should be thrown instead.

Fixes:
 - Rename all mispelled variables to the correct ones.
 - Add the correct error when the the api secret is missing.